backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
notion-client
python-dotenv
requests
google-auth-oauthlib
google-api-python-client
